+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ## [Unreleased]
 
+## [3.1.0] - 2022-06-22
+
 ### Changed
 
 - improved error message for database checks on folder paths